+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 
 ### Features
 - Optical parameters for sensor tables are now loaded from values entered in "[default_sensor_values]" in config.toml
-- Can now change TIR and SPR angle fitting parameters and refit them using a button hidden menu in the Response quantification tab. NOTE: This will change the TIR and SPR fitting for hte whole session.
+- Can now change TIR and SPR angle fitting parameters and refit them using inputs and a button in a collapsable hidden menu in the Response quantification tab. NOTE: This will change the TIR and SPR fitting for hte whole session.
 - Default TIR and SPR fitting parameters added to config.toml
 - Displays the TIR and SPR angle fits in the reflectivity window when hovering over data points in the SPR sensorgram
 - A custom folder location for the "SPRpy sessions" folder and for the data prompt folder can now be set in config.toml